Repository e service para perfil

+++ src/main/java/com/devprojects/sistemaautenticacao/entities/Perfil.java
@@ -14,23 +14,23 @@ public class Perfil implements Serializable{
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String descricao;
 	
 	public Perfil() {
 	}
 
-	public Perfil(Long id, String descricao) {
+	public Perfil(Integer id, String descricao) {
 		super();
 		this.id = id;
 		this.descricao = descricao;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
